part 4 complete except ex 4.23 (fix tests)

+++ controllers/blogs.js
@@ -1,32 +1,56 @@
 const blogsRouter = require('express').Router();
-const app = require('../app');
 const Blog = require('../models/blog');
-const { blogs } = require('../utils/list_helper');
+const User = require('../models/user');
 
+const jwt = require('jsonwebtoken')
 
-blogsRouter.get('/', (request, response) => {
-  Blog
-    .find({})
-    .then(blogs => {
-      response.json(blogs)
-    })
+blogsRouter.get('/', async (request, response) => {
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
+  console.log(blogs)
+  response.json(blogs)
 })
 
 blogsRouter.post('/', async (request, response, next) => {
-  const blog = new Blog(request.body)
+  const body = request.body
 
-  if (!blog.likes) blog.likes = 0
-  if (!blog.title && !blog.url) {
-    return response.status(400).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
   }
+  // const user = await User.findById(decodedToken.id)
+  const user = request.user
+
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes,
+    user: user._id,
+  })
 
   const savedBlog = await blog.save()
-  response.status(201).json(savedBlog)
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
+  response.json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end();
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  const blogUserId = blog.user.toString()
+  const user = response.user
+
+  if (blogUserId === decodedToken.id) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end();
+  } else {
+    response.status(401).json({ error: 'only users who created the blog can delete it' })
+  }
 })
